Refactor inject parent in SDModalDialog
Getting a reference to the parent via a global variable creates
an implicit dependency. "Explicit is better than implicit."

Note: I'm deciding that the review of the two tests that I touched
is out of scope. I believe that asserting on a constructor call is
unlikely to be very useful, but I'd need to look more closely into
it to be sure.