Ensure concurrent requests are not misinterpreted (UNCOMPLETE)
Given that the export service cannot currently associate
a job ID with a request, there is no way to distinguish between
responses if concurrent jobs were to relate to different jobs.

As a result, we must limit ourselves to only supporting one
printing job at a time, or expand on the export service API
to allow for a job ID / request ID to be sent and returned.

That ID could and should remain opaque to the export service.

NOTE: This commit is not enough in itself. We also MUST ensure
that all actions to print conversations are disabled while any
printing operation is in progress.